Introduced to kNN algorithm
k Nearest Neighbors algorithms is an instance based learning algorithm. You base your decision with the k nearest neighbors. 
Thinking this as a an n-dimensional space, where n is the number of features.
We are using the Fashion MNIST dataset, where there are 10 possible classification, going from a shirt to a sneaker.